@@ -6,38 +6,42 @@
 #include "utils.h"
 
 int APIENTRY wWinMain(_In_ HINSTANCE instance, _In_opt_ HINSTANCE prev,
-                      _In_ wchar_t *command_line, _In_ int show_command) {
-  // Attach to console when present (e.g., 'flutter run') or create a
-  // new console when running with a debugger.
-  if (!::AttachConsole(ATTACH_PARENT_PROCESS) && ::IsDebuggerPresent()) {
-    CreateAndAttachConsole();
-  }
+                      _In_ wchar_t *command_line, _In_ int show_command)
+{
+    // Attach to console when present (e.g., 'flutter run') or create a
+    // new console when running with a debugger.
+    if (!::AttachConsole(ATTACH_PARENT_PROCESS) && ::IsDebuggerPresent())
+    {
+        CreateAndAttachConsole();
+    }
 
-  // Initialize COM, so that it is available for use in the library and/or
-  // plugins.
-  ::CoInitializeEx(nullptr, COINIT_APARTMENTTHREADED);
+    // Initialize COM, so that it is available for use in the library and/or
+    // plugins.
+    ::CoInitializeEx(nullptr, COINIT_APARTMENTTHREADED);
 
-  flutter::DartProject project(L"data");
+    flutter::DartProject project(L"data");
 
-  std::vector<std::string> command_line_arguments =
-      GetCommandLineArguments();
+    std::vector<std::string> command_line_arguments =
+        GetCommandLineArguments();
 
-  project.set_dart_entrypoint_arguments(std::move(command_line_arguments));
+    project.set_dart_entrypoint_arguments(std::move(command_line_arguments));
 
-  FlutterWindow window(project);
-  Win32Window::Point origin(10, 10);
-  Win32Window::Size size(1280, 720);
-  if (!window.Create(L"fast_calc", origin, size)) {
-    return EXIT_FAILURE;
-  }
-  window.SetQuitOnClose(true);
+    FlutterWindow window(project);
+    Win32Window::Point origin(10, 10);
+    Win32Window::Size size(1280, 720);
+    if (!window.Create(L"flutter_application_1", origin, size))
+    {
+        return EXIT_FAILURE;
+    }
+    window.SetQuitOnClose(true);
 
-  ::MSG msg;
-  while (::GetMessage(&msg, nullptr, 0, 0)) {
-    ::TranslateMessage(&msg);
-    ::DispatchMessage(&msg);
-  }
+    ::MSG msg;
+    while (::GetMessage(&msg, nullptr, 0, 0))
+    {
+        ::TranslateMessage(&msg);
+        ::DispatchMessage(&msg);
+    }
 
-  ::CoUninitialize();
-  return EXIT_SUCCESS;
+    ::CoUninitialize();
+    return EXIT_SUCCESS;
 }
